- moved replacer plugin methods to own class "ReplacerPluginConfigurator" - configured replacerPlugin with replacer rules before recalling targets from sitemap
